+++ src/inspectors/CollaborationInspector.spec.ts
@@ -33,7 +33,14 @@ describe('Collaboration Inspector', () => {
   it('returns paginated pull requests', async () => {
     new GitHubNock('1', 'octocat', 1296269, 'Hello-World').getPulls({
       pulls: [
-        { number: 1, state: 'open', title: 'new-feature', body: 'Please pull these awesome changes', head: 'new-topic', base: 'master' },
+        {
+          number: 1,
+          state: 'open',
+          title: 'Edited README via GitHub',
+          body: 'Please pull these awesome changes',
+          head: 'new-topic',
+          base: 'master',
+        },
       ],
     });