+++ src/practices/IPractice.ts
@@ -3,7 +3,7 @@ import { PracticeContext } from '../contexts/practice/PracticeContext';
 import { ReportTable, ReportText } from '../reporters/ReporterData';
 import { FixerContext } from '../contexts/fixer/FixerContext';
 import { SecurityIssueDto } from '..';
-import { SecurityIssueSummaryDto, UpdatedDependencyDto, LinterIssueDto } from '../reporters';
+import { SecurityIssueSummaryDto, UpdatedDependencyDto, LinterIssueDto, PullRequestDto } from '../reporters';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export interface IPractice<T extends {} = {}> {
@@ -49,4 +49,5 @@ export type Statistics = {
   };
   updatedDependencies?: UpdatedDependencyDto[];
   linterIssues?: LinterIssueDto[];
+  pullRequests?: PullRequestDto[];
 };